Rename filter set workplace -> workspace
